requirements.txt
# --- Core Web & HTML Tools ---
requests
beautifulsoup4
lxml
html5lib

# --- NLP & Scoring ---
scikit-learn
numpy
pandas
sentence-transformers  # For embeddings

# --- LLM Integration (GPT-4o) ---
openai

# --- Database ---
sqlalchemy
python-dotenv

# --- API ---
fastapi
uvicorn[standard]
pydantic

# --- Markdown & Report Generation ---
markdown2
Jinja2
pdfkit  # switched from weasyprint
# wkhtmltopdf must be installed separately (binary dependency)

# --- Streamlit UI (Optional but safe to include) ---
streamlit

# --- Misc Tools ---
tqdm  # progress bars (optional but useful)
